+++ src/core/export_pipeline/implementation/media_postprocess.rs
@@ -1,32 +1,32 @@
-use std::collections::{HashMap, VecDeque};
-use std::io::{Cursor, Read, Seek};
+//! Post-processing a bundle's exported files: media conversion.
+//!
+//! This file orchestrates; the submodules do the work.
+
+pub(super) mod acb;
+pub(super) mod encode_slots;
+mod hca;
+mod model;
+mod timing;
+pub(super) mod usm;
+
 use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::{Arc, Condvar, Mutex, OnceLock};
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use super::images::handle_png_conversion;
 
-use crate::core::cleanup::remove_file_if_exists;
-use crate::core::codec;
-use crate::core::config::{AppConfig, AudioOutputFormat, MediaBackend, RegionConfig};
+use crate::core::config::{AppConfig, MediaBackend, RegionConfig};
 use crate::core::errors::ExportPipelineError;
-use crate::core::media::{
-    convert_hca_bytes_to_flac_with_backend, convert_hca_bytes_to_mp3_with_backend,
-    convert_m2v_bytes_to_mp4_with_backend, convert_m2v_to_mp4_with_backend,
-    convert_usm_to_mp4_with_backend, convert_wav_bytes_to_flac_with_backend,
-    convert_wav_bytes_to_mp3_with_backend, FrameRate,
-};
 
+use self::acb::handle_acb_files_owned;
+use self::model::OwnedAcbPostProcessOptions;
+use self::timing::{merge_raw_phase_ms, record_phase_ms};
+use self::usm::{process_usm_input_with_metrics, UsmPostProcessOutput};
 use super::images::convert_native_surrogate_images_to_png;
-use super::limits::{
-    acquire_cpu_budget_permit_blocking, configure_cpu_budget_throttle, CpuBudgetPermit,
-};
+use super::limits::configure_cpu_budget_throttle;
 use super::record_max_phase_ms;
 use super::tasks::{
-    merge_usm_inputs, panic_message, post_process_files_by_extension,
-    prepare_usm_processing_inputs, remove_export_file_if_exists, run_tasks, scan_all_files,
-    UsmProcessingInput,
+    merge_usm_inputs, post_process_files_by_extension, prepare_usm_processing_inputs, run_tasks,
+    scan_all_files, UsmProcessingInput,
 };
 use super::types::{NativeInMemoryMediaSource, PostProcessSummary};
 
@@ -222,138 +222,6 @@ pub(super) fn scoped_upload_files(
     files
 }
 
-pub(super) fn record_phase_ms(target: &mut HashMap<String, u64>, phase: &str, started: Instant) {
-    target.insert(
-        phase.to_string(),
-        started.elapsed().as_millis().min(u128::from(u64::MAX)) as u64,
-    );
-}
-
-pub(super) fn add_elapsed_phase_ms(
-    target: &mut HashMap<String, u64>,
-    phase: &str,
-    started: Instant,
-) {
-    add_phase_ms(
-        target,
-        phase,
-        started.elapsed().as_millis().min(u128::from(u64::MAX)) as u64,
-    );
-}
-
-pub(super) fn add_phase_ms(target: &mut HashMap<String, u64>, phase: &str, elapsed_ms: u64) {
-    *target.entry(phase.to_string()).or_default() += elapsed_ms;
-}
-
-pub(super) fn merge_raw_phase_ms(target: &mut HashMap<String, u64>, source: &HashMap<String, u64>) {
-    for (key, value) in source {
-        *target.entry(key.clone()).or_default() += *value;
-    }
-}
-
-pub(super) struct MediaEncodeLimiter {
-    pub(super) max: usize,
-    pub(super) state: Mutex<usize>,
-    pub(super) available: Condvar,
-}
-
-type MediaEncodeLimiterKey = (MediaEncodeKind, usize);
-type MediaEncodeLimiterMap = HashMap<MediaEncodeLimiterKey, Arc<MediaEncodeLimiter>>;
-
-pub(super) struct MediaEncodePermit {
-    pub(super) limiter: Arc<MediaEncodeLimiter>,
-}
-
-impl Drop for MediaEncodePermit {
-    fn drop(&mut self) {
-        let mut active = self.limiter.state.lock().unwrap();
-        *active = active.saturating_sub(1);
-        self.limiter.available.notify_one();
-    }
-}
-
-pub(super) struct MediaEncodeAcquire {
-    pub(super) permit: MediaEncodePermit,
-    pub(super) cpu_permit: CpuBudgetPermit,
-    pub(super) kind: MediaEncodeKind,
-    pub(super) wait_ms: u64,
-    pub(super) cpu_budget_wait_ms: u64,
-    pub(super) active: usize,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub(super) enum MediaEncodeKind {
-    Audio,
-    Video,
-}
-
-impl MediaEncodeKind {
-    fn as_metric_prefix(self) -> &'static str {
-        match self {
-            Self::Audio => "audio_encode",
-            Self::Video => "video_encode",
-        }
-    }
-}
-
-pub(super) fn acquire_media_encode_permit(
-    kind: MediaEncodeKind,
-    concurrency: usize,
-    cpu_budget: usize,
-) -> Result<MediaEncodeAcquire, ExportPipelineError> {
-    let limiter = media_encode_limiter(kind, concurrency);
-    let wait_started = Instant::now();
-    let mut active = limiter.state.lock().unwrap();
-    while *active >= limiter.max {
-        active = limiter.available.wait(active).unwrap();
-    }
-    *active += 1;
-    let active_count = *active;
-    drop(active);
-    let cpu_slot = acquire_cpu_budget_permit_blocking(cpu_budget)?;
-    Ok(MediaEncodeAcquire {
-        permit: MediaEncodePermit { limiter },
-        cpu_permit: cpu_slot.permit,
-        kind,
-        wait_ms: wait_started.elapsed().as_millis().min(u128::from(u64::MAX)) as u64,
-        cpu_budget_wait_ms: cpu_slot.wait_ms,
-        active: active_count,
-    })
-}
-
-pub(super) async fn acquire_media_encode_permit_async(
-    kind: MediaEncodeKind,
-    concurrency: usize,
-    cpu_budget: usize,
-) -> Result<MediaEncodeAcquire, ExportPipelineError> {
-    tokio::task::spawn_blocking(move || acquire_media_encode_permit(kind, concurrency, cpu_budget))
-        .await
-        .map_err(|source| ExportPipelineError::WorkerPanic {
-            worker: format!("{} limiter", kind.as_metric_prefix()),
-            message: source.to_string(),
-        })?
-}
-
-pub(super) fn media_encode_limiter(
-    kind: MediaEncodeKind,
-    concurrency: usize,
-) -> Arc<MediaEncodeLimiter> {
-    let concurrency = concurrency.max(1);
-    static LIMITERS: OnceLock<Mutex<MediaEncodeLimiterMap>> = OnceLock::new();
-    let limiters = LIMITERS.get_or_init(|| Mutex::new(HashMap::new()));
-    let mut limiters = limiters.lock().unwrap();
-    limiters
-        .entry((kind, concurrency))
-        .or_insert_with(|| {
-            Arc::new(MediaEncodeLimiter {
-                max: concurrency,
-                state: Mutex::new(0),
-                available: Condvar::new(),
-            })
-        })
-        .clone()
-}
-
 #[allow(clippy::too_many_arguments)]
 pub(super) async fn handle_usm_files(
     export_path: &Path,
@@ -490,1614 +358,4 @@ pub(super) async fn handle_usm_files(
     output.generated_files.extend(file_output.generated_files);
     merge_raw_phase_ms(&mut output.phase_ms, &file_output.phase_ms);
     Ok(output)
-}
-
-#[cfg(test)]
-#[allow(clippy::too_many_arguments)]
-pub(super) async fn process_usm_file(
-    usm_file: &Path,
-    export_path: &Path,
-    region: &RegionConfig,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-) -> Result<Vec<PathBuf>, ExportPipelineError> {
-    Ok(process_usm_input_with_metrics(
-        &UsmProcessingInput::Path(usm_file.to_path_buf()),
-        export_path,
-        region,
-        ffmpeg_path,
-        media_backend,
-        retry,
-        video_encode_concurrency,
-        cpu_budget,
-    )
-    .await?
-    .generated_files)
-}
-
-#[derive(Debug, Default)]
-pub(super) struct UsmPostProcessOutput {
-    pub(super) generated_files: Vec<PathBuf>,
-    pub(super) phase_ms: HashMap<String, u64>,
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(super) async fn process_usm_input_with_metrics(
-    usm_input: &UsmProcessingInput,
-    export_path: &Path,
-    region: &RegionConfig,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-) -> Result<UsmPostProcessOutput, ExportPipelineError> {
-    let mut output = UsmPostProcessOutput::default();
-    let output_name = usm_input.output_name()?;
-    let writes_mp4 = region.export.video.writes_mp4();
-    let writes_m2v = region.export.video.writes_m2v();
-
-    if skip_invalid_usm_input(usm_input, &mut output)? {
-        return Ok(output);
-    }
-
-    if process_direct_usm_path(
-        usm_input,
-        export_path,
-        &output_name,
-        region,
-        ffmpeg_path,
-        media_backend,
-        retry,
-        video_encode_concurrency,
-        cpu_budget,
-        &mut output,
-    )
-    .await?
-    {
-        return Ok(output);
-    }
-
-    let frame_rate = usm_frame_rate(usm_input);
-
-    if process_video_only_usm(
-        usm_input,
-        export_path,
-        &output_name,
-        writes_mp4,
-        writes_m2v,
-        ffmpeg_path,
-        media_backend,
-        frame_rate,
-        retry,
-        video_encode_concurrency,
-        cpu_budget,
-        &mut output,
-    )
-    .await?
-    {
-        return Ok(output);
-    }
-
-    if matches!(usm_input, UsmProcessingInput::Bytes { .. }) {
-        process_memory_usm(
-            usm_input,
-            export_path,
-            &output_name,
-            writes_mp4,
-            writes_m2v,
-            ffmpeg_path,
-            media_backend,
-            frame_rate,
-            retry,
-            video_encode_concurrency,
-            cpu_budget,
-            &mut output,
-        )
-        .await?;
-        return Ok(output);
-    }
-
-    process_path_usm(
-        usm_input,
-        export_path,
-        &output_name,
-        writes_mp4,
-        writes_m2v,
-        ffmpeg_path,
-        media_backend,
-        frame_rate,
-        retry,
-        video_encode_concurrency,
-        cpu_budget,
-        &mut output,
-    )
-    .await?;
-    Ok(output)
-}
-
-fn skip_invalid_usm_input(
-    usm_input: &UsmProcessingInput,
-    output: &mut UsmPostProcessOutput,
-) -> Result<bool, ExportPipelineError> {
-    if usm_input_has_crid_magic(usm_input)? {
-        return Ok(false);
-    }
-    if let Some(usm_file) = usm_input.path() {
-        tracing::warn!(
-            path = %usm_file.display(),
-            "skipping .usm post-process input without CRID magic"
-        );
-        output.generated_files.push(usm_file.to_path_buf());
-    } else {
-        tracing::warn!("skipping in-memory .usm post-process input without CRID magic");
-        usm_input.cleanup_sources()?;
-    }
-    Ok(true)
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn process_direct_usm_path(
-    usm_input: &UsmProcessingInput,
-    export_path: &Path,
-    output_name: &str,
-    region: &RegionConfig,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-    output: &mut UsmPostProcessOutput,
-) -> Result<bool, ExportPipelineError> {
-    if !region.export.video.writes_mp4()
-        || region.export.video.writes_m2v()
-        || !region.export.video.direct_mp4
-    {
-        return Ok(false);
-    }
-    let Some(usm_file) = usm_input.path() else {
-        return Ok(false);
-    };
-    let mp4 = export_path.join(format!("{output_name}.mp4"));
-    let encode_slot = acquire_media_encode_permit_async(
-        MediaEncodeKind::Video,
-        video_encode_concurrency,
-        cpu_budget,
-    )
-    .await?;
-    record_usm_video_encode_acquire(&mut output.phase_ms, &encode_slot);
-    let phase_started = Instant::now();
-    convert_usm_to_mp4_with_backend(usm_file, &mp4, ffmpeg_path, media_backend, retry).await?;
-    drop(encode_slot.cpu_permit);
-    drop(encode_slot.permit);
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.usm.convert_mp4",
-        phase_started,
-    );
-    usm_input.cleanup_sources()?;
-    output.generated_files.push(mp4);
-    Ok(true)
-}
-
-fn usm_frame_rate(usm_input: &UsmProcessingInput) -> Option<FrameRate> {
-    let UsmProcessingInput::Path(usm_file) = usm_input else {
-        return None;
-    };
-    codec::read_usm_metadata(usm_file)
-        .ok()
-        .as_ref()
-        .and_then(|metadata| metadata.video_frame_rate())
-        .filter(|(_, denominator)| *denominator > 0)
-        .map(FrameRate::from_tuple)
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn process_video_only_usm(
-    usm_input: &UsmProcessingInput,
-    export_path: &Path,
-    output_name: &str,
-    writes_mp4: bool,
-    writes_m2v: bool,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    frame_rate: Option<FrameRate>,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-    output: &mut UsmPostProcessOutput,
-) -> Result<bool, ExportPipelineError> {
-    if !writes_mp4 || writes_m2v {
-        return Ok(false);
-    }
-    let phase_started = Instant::now();
-    let streams = export_usm_input_to_memory(usm_input, false)?;
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.usm.extract",
-        phase_started,
-    );
-    let Some(video) = streams
-        .into_iter()
-        .find(|stream| stream.extension.eq_ignore_ascii_case("m2v"))
-    else {
-        return Ok(false);
-    };
-    let mp4 = export_path.join(format!("{output_name}.mp4"));
-    convert_usm_m2v_bytes(
-        &video.data,
-        &mp4,
-        ffmpeg_path,
-        media_backend,
-        frame_rate,
-        retry,
-        video_encode_concurrency,
-        cpu_budget,
-        &mut output.phase_ms,
-    )
-    .await?;
-    usm_input.cleanup_sources()?;
-    output.generated_files.push(mp4);
-    Ok(true)
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn process_memory_usm(
-    usm_input: &UsmProcessingInput,
-    export_path: &Path,
-    output_name: &str,
-    writes_mp4: bool,
-    writes_m2v: bool,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    frame_rate: Option<FrameRate>,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-    output: &mut UsmPostProcessOutput,
-) -> Result<(), ExportPipelineError> {
-    let phase_started = Instant::now();
-    let mut streams = export_usm_input_to_memory(usm_input, true)?;
-    let mut generated = write_usm_streams(export_path, &streams)?;
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.usm.extract",
-        phase_started,
-    );
-    let video = writes_mp4
-        .then(|| {
-            streams
-                .iter()
-                .position(|stream| stream.extension.eq_ignore_ascii_case("m2v"))
-                .map(|position| streams.swap_remove(position))
-        })
-        .flatten();
-    if let Some(video) = video {
-        let mp4 = export_path.join(format!("{output_name}.mp4"));
-        convert_usm_m2v_bytes(
-            &video.data,
-            &mp4,
-            ffmpeg_path,
-            media_backend,
-            frame_rate,
-            retry,
-            video_encode_concurrency,
-            cpu_budget,
-            &mut output.phase_ms,
-        )
-        .await?;
-        generated.push(mp4);
-        if !writes_m2v {
-            generated.retain(|path| !has_extension(path, "m2v"));
-            let m2v = export_path.join(format!("{}.m2v", video.name));
-            remove_file_if_exists(&m2v)
-                .map_err(|source| ExportPipelineError::Io { path: m2v, source })?;
-        }
-    }
-    usm_input.cleanup_sources()?;
-    output.generated_files = generated;
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn process_path_usm(
-    usm_input: &UsmProcessingInput,
-    export_path: &Path,
-    output_name: &str,
-    writes_mp4: bool,
-    writes_m2v: bool,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    frame_rate: Option<FrameRate>,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-    output: &mut UsmPostProcessOutput,
-) -> Result<(), ExportPipelineError> {
-    let usm_file = usm_input
-        .path()
-        .expect("non-memory USM processing requires a path");
-    let phase_started = Instant::now();
-    let extracted = codec::export_usm(usm_file, export_path)?;
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.usm.extract",
-        phase_started,
-    );
-    let mut generated = extracted.clone();
-    if writes_mp4 {
-        for extracted_file in extracted
-            .into_iter()
-            .filter(|path| has_extension(path, "m2v"))
-        {
-            let mp4 = export_path.join(format!("{output_name}.mp4"));
-            convert_usm_m2v_path(
-                &extracted_file,
-                &mp4,
-                !writes_m2v,
-                ffmpeg_path,
-                media_backend,
-                frame_rate,
-                retry,
-                video_encode_concurrency,
-                cpu_budget,
-                &mut output.phase_ms,
-            )
-            .await?;
-            generated.push(mp4);
-            if !writes_m2v {
-                generated.retain(|path| path != &extracted_file);
-            }
-        }
-    }
-    usm_input.cleanup_sources()?;
-    output.generated_files = generated;
-    Ok(())
-}
-
-fn has_extension(path: &Path, extension: &str) -> bool {
-    path.extension()
-        .and_then(|ext| ext.to_str())
-        .is_some_and(|ext| ext.eq_ignore_ascii_case(extension))
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn convert_usm_m2v_bytes(
-    video: &[u8],
-    mp4: &Path,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    frame_rate: Option<FrameRate>,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-    phase_ms: &mut HashMap<String, u64>,
-) -> Result<(), ExportPipelineError> {
-    let encode_slot = acquire_media_encode_permit_async(
-        MediaEncodeKind::Video,
-        video_encode_concurrency,
-        cpu_budget,
-    )
-    .await?;
-    record_usm_video_encode_acquire(phase_ms, &encode_slot);
-    let phase_started = Instant::now();
-    convert_m2v_bytes_to_mp4_with_backend(
-        video,
-        mp4,
-        ffmpeg_path,
-        media_backend,
-        frame_rate,
-        retry,
-    )
-    .await?;
-    drop(encode_slot.cpu_permit);
-    drop(encode_slot.permit);
-    add_elapsed_phase_ms(phase_ms, "post_process.usm.convert_mp4", phase_started);
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn convert_usm_m2v_path(
-    m2v: &Path,
-    mp4: &Path,
-    remove_source: bool,
-    ffmpeg_path: &str,
-    media_backend: MediaBackend,
-    frame_rate: Option<FrameRate>,
-    retry: &crate::core::config::RetryConfig,
-    video_encode_concurrency: usize,
-    cpu_budget: usize,
-    phase_ms: &mut HashMap<String, u64>,
-) -> Result<(), ExportPipelineError> {
-    let encode_slot = acquire_media_encode_permit_async(
-        MediaEncodeKind::Video,
-        video_encode_concurrency,
-        cpu_budget,
-    )
-    .await?;
-    record_usm_video_encode_acquire(phase_ms, &encode_slot);
-    let phase_started = Instant::now();
-    convert_m2v_to_mp4_with_backend(
-        m2v,
-        mp4,
-        remove_source,
-        ffmpeg_path,
-        media_backend,
-        frame_rate,
-        retry,
-    )
-    .await?;
-    drop(encode_slot.cpu_permit);
-    drop(encode_slot.permit);
-    add_elapsed_phase_ms(phase_ms, "post_process.usm.convert_mp4", phase_started);
-    Ok(())
-}
-
-pub(super) fn usm_input_has_crid_magic(
-    usm_input: &UsmProcessingInput,
-) -> Result<bool, ExportPipelineError> {
-    match usm_input {
-        UsmProcessingInput::Path(usm_file) => {
-            codec::file_has_usm_magic(usm_file).map_err(ExportPipelineError::from)
-        }
-        UsmProcessingInput::Bytes { data, .. } => Ok(codec::has_usm_magic(data)),
-    }
-}
-
-pub(super) fn export_usm_input_to_memory(
-    usm_input: &UsmProcessingInput,
-    export_audio: bool,
-) -> Result<Vec<cridecoder::ExtractedUsmStream>, ExportPipelineError> {
-    match usm_input {
-        UsmProcessingInput::Path(usm_file) => {
-            let usm_reader =
-                std::fs::File::open(usm_file).map_err(|source| ExportPipelineError::Io {
-                    path: usm_file.to_path_buf(),
-                    source,
-                })?;
-            let fallback_name = usm_file
-                .file_name()
-                .and_then(|name| name.to_str())
-                .unwrap_or("input.usm");
-            codec::export_usm_reader_to_memory(usm_reader, fallback_name.as_bytes(), export_audio)
-                .map_err(ExportPipelineError::from)
-        }
-        UsmProcessingInput::Bytes {
-            fallback_name,
-            data,
-            ..
-        } => codec::export_usm_to_memory(data, fallback_name.as_bytes(), export_audio)
-            .map_err(ExportPipelineError::from),
-    }
-}
-
-pub(super) fn write_usm_streams(
-    export_path: &Path,
-    streams: &[cridecoder::ExtractedUsmStream],
-) -> Result<Vec<PathBuf>, ExportPipelineError> {
-    let mut generated = Vec::with_capacity(streams.len());
-    for stream in streams {
-        let path = export_path.join(format!("{}.{}", stream.name, stream.extension));
-        std::fs::write(&path, &stream.data).map_err(|source| ExportPipelineError::Io {
-            path: path.clone(),
-            source,
-        })?;
-        generated.push(path);
-    }
-    Ok(generated)
-}
-
-pub(super) fn handle_acb_files_owned(
-    options: &OwnedAcbPostProcessOptions,
-    acb_concurrency: usize,
-    scoped_post_process: bool,
-    scoped_files: &[PathBuf],
-    acb_sources: Vec<NativeInMemoryMediaSource>,
-) -> Result<AcbPostProcessOutput, ExportPipelineError> {
-    let borrowed = AcbPostProcessOptions {
-        output_dir: &options.output_dir,
-        region: &options.region,
-        ffmpeg_path: &options.ffmpeg_path,
-        media_backend: options.media_backend,
-        retry: &options.retry,
-        hca_concurrency: options.hca_concurrency,
-        audio_encode_concurrency: options.audio_encode_concurrency,
-        cpu_budget: options.cpu_budget,
-    };
-    handle_acb_files(
-        &borrowed,
-        acb_concurrency,
-        scoped_post_process,
-        scoped_files,
-        acb_sources,
-    )
-}
-
-pub(super) fn handle_acb_files(
-    options: &AcbPostProcessOptions<'_>,
-    acb_concurrency: usize,
-    scoped_post_process: bool,
-    scoped_files: &[PathBuf],
-    acb_sources: Vec<NativeInMemoryMediaSource>,
-) -> Result<AcbPostProcessOutput, ExportPipelineError> {
-    let acb_files = post_process_files_by_extension(
-        options.output_dir,
-        scoped_post_process,
-        scoped_files,
-        "acb",
-    )?;
-    if !options.region.export.acb.export
-        || !options.region.export.acb.decode
-        || (acb_files.is_empty() && acb_sources.is_empty())
-    {
-        return Ok(AcbPostProcessOutput::default());
-    }
-
-    if acb_files.len() + acb_sources.len() == 1 || !options.region.export.hca.decode {
-        return handle_acb_files_batched(acb_files, acb_sources, options, acb_concurrency);
-    }
-    handle_acb_files_streaming(acb_files, acb_sources, options, acb_concurrency)
-}
-
-pub(super) fn handle_acb_files_batched(
-    acb_files: Vec<PathBuf>,
-    acb_sources: Vec<NativeInMemoryMediaSource>,
-    options: &AcbPostProcessOptions<'_>,
-    acb_concurrency: usize,
-) -> Result<AcbPostProcessOutput, ExportPipelineError> {
-    let acb_inputs = acb_extraction_inputs(acb_files, acb_sources);
-    let acb_file_count = acb_inputs.len();
-    let output_dir = options.output_dir.to_path_buf();
-    let region = options.region.clone();
-    let ffmpeg_path = options.ffmpeg_path.to_string();
-    let retry = options.retry.clone();
-    let media_backend = options.media_backend;
-    let hca_concurrency = options.hca_concurrency;
-    let audio_encode_concurrency = options.audio_encode_concurrency;
-    let cpu_budget = options.cpu_budget;
-    let extracted = run_tasks(acb_inputs, acb_concurrency, move |acb_input| {
-        let options = AcbPostProcessOptions {
-            output_dir: &output_dir,
-            region: &region,
-            ffmpeg_path: &ffmpeg_path,
-            media_backend,
-            retry: &retry,
-            hca_concurrency,
-            audio_encode_concurrency,
-            cpu_budget,
-        };
-        extract_acb_tracks_from_input(acb_input, &options)
-    })?;
-    let mut merged = AcbPostProcessOutput::default();
-    merged.phase_ms.insert(
-        "media_scheduler.acb_file_count".to_string(),
-        acb_file_count as u64,
-    );
-    merged.phase_ms.insert(
-        "media_scheduler.acb_worker_count".to_string(),
-        acb_concurrency.max(1).min(acb_file_count) as u64,
-    );
-    let mut hca_tracks = Vec::new();
-    let mut source_files = Vec::new();
-    for output in extracted {
-        merged.generated_files.extend(output.generated_files);
-        merge_raw_phase_ms(&mut merged.phase_ms, &output.phase_ms);
-        let track_output_dir = output.output_dir.clone();
-        hca_tracks.extend(
-            output
-                .hca_tracks
-                .into_iter()
-                .map(|track| HcaTrackProcessJob {
-                    track,
-                    output_dir: track_output_dir.clone(),
-                }),
-        );
-        if let Some(source_file) = output.source_file {
-            source_files.push(source_file);
-        }
-    }
-
-    let phase_started = Instant::now();
-    let hca_output = process_hca_tracks(hca_tracks, options)?;
-    merged.generated_files.extend(hca_output.generated_files);
-    merge_raw_phase_ms(&mut merged.phase_ms, &hca_output.phase_ms);
-    add_elapsed_phase_ms(
-        &mut merged.phase_ms,
-        "post_process.acb.hca_tracks_wall",
-        phase_started,
-    );
-
-    for source_file in source_files {
-        let phase_started = Instant::now();
-        remove_export_file_if_exists(&source_file)?;
-        add_elapsed_phase_ms(
-            &mut merged.phase_ms,
-            "post_process.acb.remove_source",
-            phase_started,
-        );
-    }
-    Ok(merged)
-}
-
-pub(super) fn handle_acb_files_streaming(
-    acb_files: Vec<PathBuf>,
-    acb_sources: Vec<NativeInMemoryMediaSource>,
-    options: &AcbPostProcessOptions<'_>,
-    acb_concurrency: usize,
-) -> Result<AcbPostProcessOutput, ExportPipelineError> {
-    let acb_inputs = acb_extraction_inputs(acb_files, acb_sources);
-    let acb_file_count = acb_inputs.len();
-    let acb_worker_count = acb_concurrency.max(1).min(acb_file_count);
-    let hca_worker_count = options.hca_concurrency.max(1);
-    let queue_capacity = hca_worker_count.saturating_mul(2).max(1);
-    let (track_sender, track_receiver) =
-        std::sync::mpsc::sync_channel::<HcaTrackProcessJob>(queue_capacity);
-    let track_receiver = Arc::new(Mutex::new(track_receiver));
-    let state = AcbStreamingState::default();
-    let hca_started = Instant::now();
-    let hca_handles =
-        spawn_hca_stream_workers(hca_worker_count, track_receiver, state.clone(), options)?;
-
-    let acb_queue = Arc::new(Mutex::new(VecDeque::from(acb_inputs)));
-    let acb_handles = spawn_acb_stream_workers(
-        acb_worker_count,
-        acb_queue,
-        track_sender.clone(),
-        state.clone(),
-        options,
-    )?;
-    drop(track_sender);
-    join_stream_workers(acb_handles, "acb track extract")?;
-    join_stream_workers(hca_handles, "hca memory export")?;
-    finish_acb_streaming(
-        state,
-        acb_file_count,
-        acb_worker_count,
-        hca_worker_count,
-        hca_started,
-    )
-}
-
-#[derive(Clone, Default)]
-struct AcbStreamingState {
-    results: Arc<Mutex<Vec<PathBuf>>>,
-    phase_ms: Arc<Mutex<HashMap<String, u64>>>,
-    source_files: Arc<Mutex<Vec<PathBuf>>>,
-    first_error: Arc<Mutex<Option<ExportPipelineError>>>,
-    hca_track_count: Arc<AtomicUsize>,
-}
-
-fn spawn_hca_stream_workers(
-    worker_count: usize,
-    receiver: Arc<Mutex<std::sync::mpsc::Receiver<HcaTrackProcessJob>>>,
-    state: AcbStreamingState,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<Vec<std::thread::JoinHandle<()>>, ExportPipelineError> {
-    let mut handles = Vec::with_capacity(worker_count);
-    for _ in 0..worker_count {
-        let receiver = receiver.clone();
-        let state = state.clone();
-        let owned_options = OwnedAcbPostProcessOptions::from(options);
-        handles.push(
-            std::thread::Builder::new()
-                .name("hca-memory-export".to_string())
-                .stack_size(32 * 1024 * 1024)
-                .spawn(move || run_hca_stream_worker(receiver, state, owned_options))
-                .map_err(|source| ExportPipelineError::Io {
-                    path: options.output_dir.to_path_buf(),
-                    source,
-                })?,
-        );
-    }
-    Ok(handles)
-}
-
-fn run_hca_stream_worker(
-    receiver: Arc<Mutex<std::sync::mpsc::Receiver<HcaTrackProcessJob>>>,
-    state: AcbStreamingState,
-    options: OwnedAcbPostProcessOptions,
-) {
-    while state.first_error.lock().unwrap().is_none() {
-        let Ok(job) = receiver.lock().unwrap().recv() else {
-            break;
-        };
-        let track_options = HcaTrackProcessOptions {
-            output_dir: &job.output_dir,
-            region: &options.region,
-            ffmpeg_path: &options.ffmpeg_path,
-            media_backend: options.media_backend,
-            retry: &options.retry,
-            audio_encode_concurrency: options.audio_encode_concurrency,
-            cpu_budget: options.cpu_budget,
-        };
-        match process_hca_track(job.track, &track_options) {
-            Ok(output) => merge_hca_stream_output(&state, output),
-            Err(err) => {
-                set_first_error(&state.first_error, err);
-                break;
-            }
-        }
-    }
-}
-
-fn merge_hca_stream_output(state: &AcbStreamingState, output: HcaTrackProcessOutput) {
-    state.results.lock().unwrap().extend(output.generated_files);
-    merge_raw_phase_ms(&mut state.phase_ms.lock().unwrap(), &output.phase_ms);
-}
-
-fn spawn_acb_stream_workers(
-    worker_count: usize,
-    queue: Arc<Mutex<VecDeque<AcbExtractionInput>>>,
-    sender: std::sync::mpsc::SyncSender<HcaTrackProcessJob>,
-    state: AcbStreamingState,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<Vec<std::thread::JoinHandle<()>>, ExportPipelineError> {
-    let mut handles = Vec::with_capacity(worker_count);
-    for _ in 0..worker_count {
-        let queue = queue.clone();
-        let sender = sender.clone();
-        let state = state.clone();
-        let owned_options = OwnedAcbPostProcessOptions::from(options);
-        handles.push(
-            std::thread::Builder::new()
-                .name("acb-track-extract".to_string())
-                .stack_size(32 * 1024 * 1024)
-                .spawn(move || run_acb_stream_worker(queue, sender, state, owned_options))
-                .map_err(|source| ExportPipelineError::Io {
-                    path: options.output_dir.to_path_buf(),
-                    source,
-                })?,
-        );
-    }
-    Ok(handles)
-}
-
-fn run_acb_stream_worker(
-    queue: Arc<Mutex<VecDeque<AcbExtractionInput>>>,
-    sender: std::sync::mpsc::SyncSender<HcaTrackProcessJob>,
-    state: AcbStreamingState,
-    options: OwnedAcbPostProcessOptions,
-) {
-    while state.first_error.lock().unwrap().is_none() {
-        let Some(input) = queue.lock().unwrap().pop_front() else {
-            break;
-        };
-        let worker_options = options.as_borrowed();
-        match extract_acb_tracks_from_input(input, &worker_options) {
-            Ok(output) => forward_acb_stream_output(output, &sender, &state),
-            Err(err) => {
-                set_first_error(&state.first_error, err);
-                break;
-            }
-        }
-    }
-}
-
-fn forward_acb_stream_output(
-    output: AcbTrackExtractionOutput,
-    sender: &std::sync::mpsc::SyncSender<HcaTrackProcessJob>,
-    state: &AcbStreamingState,
-) {
-    state.results.lock().unwrap().extend(output.generated_files);
-    merge_raw_phase_ms(&mut state.phase_ms.lock().unwrap(), &output.phase_ms);
-    if let Some(source_file) = output.source_file {
-        state.source_files.lock().unwrap().push(source_file);
-    }
-    for track in output.hca_tracks {
-        state.hca_track_count.fetch_add(1, Ordering::Relaxed);
-        let job = HcaTrackProcessJob {
-            track,
-            output_dir: output.output_dir.clone(),
-        };
-        if !send_hca_track(sender, job, &state.first_error) {
-            break;
-        }
-    }
-}
-
-fn join_stream_workers(
-    handles: Vec<std::thread::JoinHandle<()>>,
-    worker: &str,
-) -> Result<(), ExportPipelineError> {
-    for handle in handles {
-        handle
-            .join()
-            .map_err(|panic| ExportPipelineError::WorkerPanic {
-                worker: worker.to_string(),
-                message: panic_message(panic),
-            })?;
-    }
-    Ok(())
-}
-
-fn finish_acb_streaming(
-    state: AcbStreamingState,
-    acb_file_count: usize,
-    acb_worker_count: usize,
-    hca_worker_count: usize,
-    hca_started: Instant,
-) -> Result<AcbPostProcessOutput, ExportPipelineError> {
-    if let Some(err) = state.first_error.lock().unwrap().take() {
-        return Err(err);
-    }
-    let mut output = AcbPostProcessOutput::default();
-    add_acb_streaming_metrics(
-        &mut output,
-        &state,
-        acb_file_count,
-        acb_worker_count,
-        hca_worker_count,
-        hca_started,
-    );
-    output.generated_files = std::mem::take(&mut *state.results.lock().unwrap());
-    remove_acb_source_files(&state.source_files, &mut output.phase_ms)?;
-    Ok(output)
-}
-
-fn add_acb_streaming_metrics(
-    output: &mut AcbPostProcessOutput,
-    state: &AcbStreamingState,
-    acb_file_count: usize,
-    acb_worker_count: usize,
-    hca_worker_count: usize,
-    hca_started: Instant,
-) {
-    for (key, value) in [
-        ("media_scheduler.acb_file_count", acb_file_count),
-        ("media_scheduler.acb_worker_count", acb_worker_count),
-        (
-            "media_scheduler.hca_track_count",
-            state.hca_track_count.load(Ordering::Relaxed),
-        ),
-        ("media_scheduler.hca_worker_count", hca_worker_count),
-    ] {
-        output.phase_ms.insert(key.to_string(), value as u64);
-    }
-    merge_raw_phase_ms(&mut output.phase_ms, &state.phase_ms.lock().unwrap());
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.acb.hca_tracks_wall",
-        hca_started,
-    );
-}
-
-fn remove_acb_source_files(
-    source_files: &Arc<Mutex<Vec<PathBuf>>>,
-    phase_ms: &mut HashMap<String, u64>,
-) -> Result<(), ExportPipelineError> {
-    for source_file in source_files.lock().unwrap().iter() {
-        let phase_started = Instant::now();
-        remove_export_file_if_exists(source_file)?;
-        add_elapsed_phase_ms(phase_ms, "post_process.acb.remove_source", phase_started);
-    }
-    Ok(())
-}
-
-pub(super) fn set_first_error(
-    first_error: &Arc<Mutex<Option<ExportPipelineError>>>,
-    err: ExportPipelineError,
-) {
-    let mut first = first_error.lock().unwrap();
-    if first.is_none() {
-        *first = Some(err);
-    }
-}
-
-pub(super) fn send_hca_track(
-    sender: &std::sync::mpsc::SyncSender<HcaTrackProcessJob>,
-    track: HcaTrackProcessJob,
-    first_error: &Arc<Mutex<Option<ExportPipelineError>>>,
-) -> bool {
-    let mut track = Some(track);
-    loop {
-        if first_error.lock().unwrap().is_some() {
-            return false;
-        }
-        match sender.try_send(track.take().expect("track is retained until sent")) {
-            Ok(()) => return true,
-            Err(std::sync::mpsc::TrySendError::Full(returned)) => {
-                track = Some(returned);
-                std::thread::sleep(Duration::from_millis(1));
-            }
-            Err(std::sync::mpsc::TrySendError::Disconnected(_)) => return false,
-        }
-    }
-}
-
-#[derive(Debug, Default)]
-pub(super) struct AcbPostProcessOutput {
-    pub(super) generated_files: Vec<PathBuf>,
-    pub(super) phase_ms: HashMap<String, u64>,
-}
-
-pub(super) struct HcaTrackProcessJob {
-    pub(super) track: SharedAcbTrack,
-    pub(super) output_dir: PathBuf,
-}
-
-#[derive(Debug, Clone)]
-pub(super) struct SharedAcbTrack {
-    pub(super) name: String,
-    pub(super) extension: String,
-    pub(super) data: Arc<[u8]>,
-}
-
-#[derive(Debug, Clone)]
-pub(super) enum AcbExtractionInput {
-    File(PathBuf),
-    Memory(NativeInMemoryMediaSource),
-}
-
-#[derive(Clone)]
-pub(super) struct OwnedAcbPostProcessOptions {
-    pub(super) output_dir: PathBuf,
-    pub(super) region: RegionConfig,
-    pub(super) ffmpeg_path: String,
-    pub(super) media_backend: MediaBackend,
-    pub(super) retry: crate::core::config::RetryConfig,
-    pub(super) hca_concurrency: usize,
-    pub(super) audio_encode_concurrency: usize,
-    pub(super) cpu_budget: usize,
-}
-
-impl OwnedAcbPostProcessOptions {
-    fn as_borrowed(&self) -> AcbPostProcessOptions<'_> {
-        AcbPostProcessOptions {
-            output_dir: &self.output_dir,
-            region: &self.region,
-            ffmpeg_path: &self.ffmpeg_path,
-            media_backend: self.media_backend,
-            retry: &self.retry,
-            hca_concurrency: self.hca_concurrency,
-            audio_encode_concurrency: self.audio_encode_concurrency,
-            cpu_budget: self.cpu_budget,
-        }
-    }
-}
-
-impl From<&AcbPostProcessOptions<'_>> for OwnedAcbPostProcessOptions {
-    fn from(options: &AcbPostProcessOptions<'_>) -> Self {
-        Self {
-            output_dir: options.output_dir.to_path_buf(),
-            region: options.region.clone(),
-            ffmpeg_path: options.ffmpeg_path.to_string(),
-            media_backend: options.media_backend,
-            retry: options.retry.clone(),
-            hca_concurrency: options.hca_concurrency,
-            audio_encode_concurrency: options.audio_encode_concurrency,
-            cpu_budget: options.cpu_budget,
-        }
-    }
-}
-
-#[derive(Clone)]
-pub(super) struct AcbPostProcessOptions<'a> {
-    pub(super) output_dir: &'a Path,
-    pub(super) region: &'a RegionConfig,
-    pub(super) ffmpeg_path: &'a str,
-    pub(super) media_backend: MediaBackend,
-    pub(super) retry: &'a crate::core::config::RetryConfig,
-    pub(super) hca_concurrency: usize,
-    pub(super) audio_encode_concurrency: usize,
-    pub(super) cpu_budget: usize,
-}
-
-#[derive(Debug, Default)]
-pub(super) struct AcbTrackExtractionOutput {
-    pub(super) hca_tracks: Vec<SharedAcbTrack>,
-    pub(super) generated_files: Vec<PathBuf>,
-    pub(super) source_file: Option<PathBuf>,
-    pub(super) output_dir: PathBuf,
-    pub(super) phase_ms: HashMap<String, u64>,
-}
-
-pub(super) fn share_acb_waveforms(
-    waveforms: Vec<cridecoder::UniqueWaveform>,
-) -> Vec<SharedAcbTrack> {
-    let mut tracks = Vec::new();
-    for waveform in waveforms {
-        let data = Arc::<[u8]>::from(waveform.data);
-        tracks.extend(waveform.cues.into_iter().map(|cue| SharedAcbTrack {
-            name: cue.name,
-            extension: waveform.extension.clone(),
-            data: data.clone(),
-        }));
-    }
-    tracks
-}
-
-pub(super) fn acb_extraction_inputs(
-    acb_files: Vec<PathBuf>,
-    acb_sources: Vec<NativeInMemoryMediaSource>,
-) -> Vec<AcbExtractionInput> {
-    acb_files
-        .into_iter()
-        .map(AcbExtractionInput::File)
-        .chain(acb_sources.into_iter().map(AcbExtractionInput::Memory))
-        .collect()
-}
-
-pub(super) fn extract_acb_tracks_from_input(
-    input: AcbExtractionInput,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<AcbTrackExtractionOutput, ExportPipelineError> {
-    match input {
-        AcbExtractionInput::File(acb_file) => extract_acb_tracks_from_file(&acb_file, options),
-        AcbExtractionInput::Memory(source) => {
-            extract_acb_tracks_from_memory_source(source, options)
-        }
-    }
-}
-
-pub(super) fn extract_acb_tracks_from_file(
-    acb_file: &Path,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<AcbTrackExtractionOutput, ExportPipelineError> {
-    let phase_started = Instant::now();
-    let acb_reader = std::fs::File::open(acb_file).map_err(|source| ExportPipelineError::Io {
-        path: acb_file.to_path_buf(),
-        source,
-    })?;
-    let open_file_ms = phase_started
-        .elapsed()
-        .as_millis()
-        .min(u128::from(u64::MAX)) as u64;
-
-    let mut output = extract_acb_tracks_from_reader(
-        acb_reader,
-        acb_file,
-        Some(acb_file.to_path_buf()),
-        options,
-    )?;
-    *output
-        .phase_ms
-        .entry("post_process.acb.open_file".to_string())
-        .or_default() += open_file_ms;
-    Ok(output)
-}
-
-pub(super) fn extract_acb_tracks_from_memory_source(
-    source: NativeInMemoryMediaSource,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<AcbTrackExtractionOutput, ExportPipelineError> {
-    extract_acb_tracks_from_reader(Cursor::new(source.payload), &source.target, None, options)
-}
-
-pub(super) fn extract_acb_tracks_from_reader<R>(
-    acb_reader: R,
-    source_hint: &Path,
-    source_file: Option<PathBuf>,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<AcbTrackExtractionOutput, ExportPipelineError>
-where
-    R: Read + Seek,
-{
-    let output_dir = source_hint
-        .parent()
-        .map(Path::to_path_buf)
-        .unwrap_or_else(|| options.output_dir.to_path_buf());
-    let mut output = AcbTrackExtractionOutput {
-        source_file,
-        output_dir,
-        ..AcbTrackExtractionOutput::default()
-    };
-
-    let hca_tracks = extract_and_filter_acb_tracks(acb_reader, source_hint, &mut output.phase_ms)?;
-
-    if !options.region.export.hca.decode {
-        return Ok(output);
-    }
-
-    output.hca_tracks = hca_tracks;
-    Ok(output)
-}
-
-fn extract_and_filter_acb_tracks<R>(
-    acb_reader: R,
-    source_hint: &Path,
-    phase_ms: &mut HashMap<String, u64>,
-) -> Result<Vec<SharedAcbTrack>, ExportPipelineError>
-where
-    R: Read + Seek,
-{
-    let phase_started = Instant::now();
-    let tracks = share_acb_waveforms(codec::export_acb_unique_to_memory(
-        acb_reader,
-        Some(source_hint),
-    )?);
-    add_elapsed_phase_ms(phase_ms, "post_process.acb.extract_tracks", phase_started);
-    let phase_started = Instant::now();
-    let tracks = filter_music_long_tracks(source_hint, tracks);
-    add_elapsed_phase_ms(phase_ms, "post_process.acb.filter_tracks", phase_started);
-    Ok(tracks)
-}
-
-fn filter_music_long_tracks(
-    source_hint: &Path,
-    mut tracks: Vec<SharedAcbTrack>,
-) -> Vec<SharedAcbTrack> {
-    let source = source_hint
-        .to_string_lossy()
-        .replace('\\', "/")
-        .to_lowercase();
-    if source.contains("music/long") {
-        tracks.retain(|track| should_keep_music_long_hca_track(&track.name, &track.extension));
-    }
-    tracks
-}
-
-pub(super) fn should_keep_music_long_hca_track(name: &str, extension: &str) -> bool {
-    let lower = format!("{name}.{extension}").to_lowercase();
-    !(lower.ends_with("_vr.hca") || lower.ends_with("_screen.hca"))
-}
-
-pub(super) fn process_hca_tracks(
-    mut hca_tracks: Vec<HcaTrackProcessJob>,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<AcbPostProcessOutput, ExportPipelineError> {
-    let mut output = AcbPostProcessOutput::default();
-    if hca_tracks.is_empty() {
-        return Ok(output);
-    }
-    output.phase_ms.insert(
-        "media_scheduler.hca_track_count".to_string(),
-        hca_tracks.len() as u64,
-    );
-
-    if hca_tracks.len() == 1 {
-        output
-            .phase_ms
-            .insert("media_scheduler.hca_worker_count".to_string(), 1);
-        let track = hca_tracks.pop().expect("single track is present");
-        let track_output = process_hca_track_job_on_large_stack(track, options)?;
-        output.generated_files.extend(track_output.generated_files);
-        merge_raw_phase_ms(&mut output.phase_ms, &track_output.phase_ms);
-        return Ok(output);
-    }
-
-    let worker_count = options.hca_concurrency.max(1).min(hca_tracks.len());
-    output.phase_ms.insert(
-        "media_scheduler.hca_worker_count".to_string(),
-        worker_count as u64,
-    );
-    let queue = Arc::new(Mutex::new(VecDeque::from(hca_tracks)));
-    let results = Arc::new(Mutex::new(Vec::<PathBuf>::new()));
-    let phase_ms = Arc::new(Mutex::new(HashMap::<String, u64>::new()));
-    let first_error = Arc::new(Mutex::new(None::<ExportPipelineError>));
-    let mut handles = Vec::with_capacity(worker_count);
-
-    for _ in 0..worker_count {
-        let queue = queue.clone();
-        let results = results.clone();
-        let phase_ms = phase_ms.clone();
-        let first_error = first_error.clone();
-        let output_dir_for_error = options.output_dir.to_path_buf();
-        let region = options.region.clone();
-        let ffmpeg_path = options.ffmpeg_path.to_string();
-        let media_backend = options.media_backend;
-        let retry = options.retry.clone();
-        let audio_encode_concurrency = options.audio_encode_concurrency;
-        let cpu_budget = options.cpu_budget;
-        let handle = std::thread::Builder::new()
-            .name("hca-memory-export".to_string())
-            .stack_size(32 * 1024 * 1024)
-            .spawn(move || loop {
-                if first_error.lock().unwrap().is_some() {
-                    break;
-                }
-
-                let next_track = queue.lock().unwrap().pop_front();
-                let Some(track_job) = next_track else {
-                    break;
-                };
-
-                let track_options = HcaTrackProcessOptions {
-                    output_dir: &track_job.output_dir,
-                    region: &region,
-                    ffmpeg_path: &ffmpeg_path,
-                    media_backend,
-                    retry: &retry,
-                    audio_encode_concurrency,
-                    cpu_budget,
-                };
-                match process_hca_track(track_job.track, &track_options) {
-                    Ok(track_output) => {
-                        results.lock().unwrap().extend(track_output.generated_files);
-                        merge_raw_phase_ms(&mut phase_ms.lock().unwrap(), &track_output.phase_ms);
-                    }
-                    Err(err) => {
-                        *first_error.lock().unwrap() = Some(err);
-                        break;
-                    }
-                }
-            })
-            .map_err(|source| ExportPipelineError::Io {
-                path: output_dir_for_error,
-                source,
-            })?;
-        handles.push(handle);
-    }
-
-    for handle in handles {
-        if let Err(payload) = handle.join() {
-            return Err(ExportPipelineError::Io {
-                path: options.output_dir.to_path_buf(),
-                source: std::io::Error::other(format!("hca worker panicked: {payload:?}")),
-            });
-        }
-    }
-
-    if let Some(err) = first_error.lock().unwrap().take() {
-        return Err(err);
-    }
-    // Workers have joined, so this holds the only reference; move the vec out instead of cloning.
-    output.generated_files = std::mem::take(&mut *results.lock().unwrap());
-    merge_raw_phase_ms(&mut output.phase_ms, &phase_ms.lock().unwrap());
-    Ok(output)
-}
-
-#[derive(Debug, Default)]
-pub(super) struct HcaTrackProcessOutput {
-    pub(super) generated_files: Vec<PathBuf>,
-    pub(super) phase_ms: HashMap<String, u64>,
-}
-
-pub(super) fn process_hca_track_job_on_large_stack(
-    track: HcaTrackProcessJob,
-    options: &AcbPostProcessOptions<'_>,
-) -> Result<HcaTrackProcessOutput, ExportPipelineError> {
-    let output_dir_for_error = track.output_dir.clone();
-    let region = options.region.clone();
-    let ffmpeg_path = options.ffmpeg_path.to_string();
-    let media_backend = options.media_backend;
-    let retry = options.retry.clone();
-    let audio_encode_concurrency = options.audio_encode_concurrency;
-    let cpu_budget = options.cpu_budget;
-    let handle = std::thread::Builder::new()
-        .name("hca-memory-export".to_string())
-        .stack_size(32 * 1024 * 1024)
-        .spawn(move || {
-            let track_options = HcaTrackProcessOptions {
-                output_dir: &track.output_dir,
-                region: &region,
-                ffmpeg_path: &ffmpeg_path,
-                media_backend,
-                retry: &retry,
-                audio_encode_concurrency,
-                cpu_budget,
-            };
-            process_hca_track(track.track, &track_options)
-        })
-        .map_err(|source| ExportPipelineError::Io {
-            path: output_dir_for_error,
-            source,
-        })?;
-    handle
-        .join()
-        .map_err(|panic| ExportPipelineError::WorkerPanic {
-            worker: "hca memory export".to_string(),
-            message: panic_message(panic),
-        })?
-}
-
-pub(super) struct HcaTrackProcessOptions<'a> {
-    pub(super) output_dir: &'a Path,
-    pub(super) region: &'a RegionConfig,
-    pub(super) ffmpeg_path: &'a str,
-    pub(super) media_backend: MediaBackend,
-    pub(super) retry: &'a crate::core::config::RetryConfig,
-    pub(super) audio_encode_concurrency: usize,
-    pub(super) cpu_budget: usize,
-}
-
-pub(super) fn record_hca_media_encode_acquire(
-    phase_ms: &mut HashMap<String, u64>,
-    encode_slot: &MediaEncodeAcquire,
-) {
-    debug_assert_eq!(encode_slot.kind, MediaEncodeKind::Audio);
-    add_phase_ms(
-        phase_ms,
-        "post_process.hca.audio_pool_wait",
-        encode_slot.wait_ms,
-    );
-    add_phase_ms(
-        phase_ms,
-        "media_scheduler.audio_encode_wait",
-        encode_slot.wait_ms,
-    );
-    record_max_phase_ms(
-        phase_ms,
-        "media_scheduler.audio_encode_active_peak",
-        encode_slot.active as u64,
-    );
-    add_phase_ms(
-        phase_ms,
-        // Compatibility metric for older bench readers.
-        "media_scheduler.media_encode_wait",
-        encode_slot.wait_ms,
-    );
-    record_max_phase_ms(
-        phase_ms,
-        "media_scheduler.media_encode_active_peak",
-        encode_slot.active as u64,
-    );
-    add_phase_ms(
-        phase_ms,
-        "media_scheduler.cpu_budget_wait",
-        encode_slot.cpu_budget_wait_ms,
-    );
-    add_phase_ms(phase_ms, "cpu_budget.wait", encode_slot.cpu_budget_wait_ms);
-}
-
-pub(super) fn record_usm_video_encode_acquire(
-    phase_ms: &mut HashMap<String, u64>,
-    encode_slot: &MediaEncodeAcquire,
-) {
-    debug_assert_eq!(encode_slot.kind, MediaEncodeKind::Video);
-    add_phase_ms(
-        phase_ms,
-        "post_process.usm.video_pool_wait",
-        encode_slot.wait_ms,
-    );
-    add_phase_ms(
-        phase_ms,
-        "media_scheduler.video_encode_wait",
-        encode_slot.wait_ms,
-    );
-    record_max_phase_ms(
-        phase_ms,
-        "media_scheduler.video_encode_active_peak",
-        encode_slot.active as u64,
-    );
-    add_phase_ms(
-        phase_ms,
-        "media_scheduler.cpu_budget_wait",
-        encode_slot.cpu_budget_wait_ms,
-    );
-    add_phase_ms(phase_ms, "cpu_budget.wait", encode_slot.cpu_budget_wait_ms);
-}
-
-pub(super) fn process_hca_track(
-    track: SharedAcbTrack,
-    options: &HcaTrackProcessOptions<'_>,
-) -> Result<HcaTrackProcessOutput, ExportPipelineError> {
-    let mut output = HcaTrackProcessOutput::default();
-    if !track.extension.eq_ignore_ascii_case("hca") {
-        return write_non_hca_track(track, options);
-    }
-
-    let Some(formats) = HcaFormatPlan::from_options(options) else {
-        return Ok(output);
-    };
-    let wav_file = options.output_dir.join(format!("{}.wav", track.name));
-
-    if formats.wav_only() {
-        decode_hca_to_wav_file(&track.data, &wav_file, options.cpu_budget, &mut output)?;
-        return Ok(output);
-    }
-
-    if !formats.encodes_audio() {
-        return Ok(output);
-    }
-
-    let wav_bytes = prepare_hca_wav_bytes(
-        &track.data,
-        &wav_file,
-        formats,
-        options.cpu_budget,
-        &mut output,
-    )?;
-
-    if formats.encode_mp3 {
-        encode_hca_mp3(&track, wav_bytes.as_deref(), options, &mut output)?;
-    }
-    if formats.encode_flac {
-        encode_hca_flac(&track, wav_bytes.as_deref(), options, &mut output)?;
-    }
-
-    Ok(output)
-}
-
-#[derive(Clone, Copy)]
-struct HcaFormatPlan {
-    keep_wav: bool,
-    encode_mp3: bool,
-    encode_flac: bool,
-}
-
-impl HcaFormatPlan {
-    fn from_options(options: &HcaTrackProcessOptions<'_>) -> Option<Self> {
-        let formats = options.region.export.audio.output_formats();
-        (!formats.is_empty()).then(|| Self {
-            keep_wav: formats.contains(&AudioOutputFormat::Wav),
-            encode_mp3: formats.contains(&AudioOutputFormat::Mp3),
-            encode_flac: formats.contains(&AudioOutputFormat::Flac),
-        })
-    }
-
-    fn wav_only(self) -> bool {
-        self.keep_wav && !self.encode_mp3 && !self.encode_flac
-    }
-
-    fn encodes_audio(self) -> bool {
-        self.encode_mp3 || self.encode_flac
-    }
-
-    fn needs_wav_bytes(self) -> bool {
-        (self.keep_wav && self.encodes_audio()) || (self.encode_mp3 && self.encode_flac)
-    }
-}
-
-fn write_non_hca_track(
-    track: SharedAcbTrack,
-    options: &HcaTrackProcessOptions<'_>,
-) -> Result<HcaTrackProcessOutput, ExportPipelineError> {
-    let mut output = HcaTrackProcessOutput::default();
-    let phase_started = Instant::now();
-    let output_path = options
-        .output_dir
-        .join(format!("{}.{}", track.name, track.extension));
-    std::fs::write(&output_path, track.data).map_err(|source| ExportPipelineError::Io {
-        path: output_path.clone(),
-        source,
-    })?;
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.hca.write_non_hca",
-        phase_started,
-    );
-    output.generated_files.push(output_path);
-    Ok(output)
-}
-
-fn record_hca_cpu_wait(output: &mut HcaTrackProcessOutput, wait_ms: u64) {
-    add_phase_ms(
-        &mut output.phase_ms,
-        "post_process.hca.cpu_budget_wait",
-        wait_ms,
-    );
-    add_phase_ms(&mut output.phase_ms, "cpu_budget.wait", wait_ms);
-}
-
-fn decode_hca_to_wav_file(
-    data: &[u8],
-    wav_file: &Path,
-    cpu_budget: usize,
-    output: &mut HcaTrackProcessOutput,
-) -> Result<(), ExportPipelineError> {
-    let cpu_slot = acquire_cpu_budget_permit_blocking(cpu_budget)?;
-    record_hca_cpu_wait(output, cpu_slot.wait_ms);
-    let phase_started = Instant::now();
-    codec::decode_hca_bytes_to_wav(data, wav_file)?;
-    drop(cpu_slot.permit);
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.hca.decode_write_wav",
-        phase_started,
-    );
-    output.generated_files.push(wav_file.to_path_buf());
-    Ok(())
-}
-
-fn prepare_hca_wav_bytes(
-    data: &[u8],
-    wav_file: &Path,
-    formats: HcaFormatPlan,
-    cpu_budget: usize,
-    output: &mut HcaTrackProcessOutput,
-) -> Result<Option<Vec<u8>>, ExportPipelineError> {
-    if !formats.needs_wav_bytes() {
-        return Ok(None);
-    }
-    let cpu_slot = acquire_cpu_budget_permit_blocking(cpu_budget)?;
-    record_hca_cpu_wait(output, cpu_slot.wait_ms);
-    let phase_started = Instant::now();
-    let wav_bytes = codec::decode_hca_bytes_to_wav_bytes(data)?;
-    drop(cpu_slot.permit);
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.hca.decode_wav",
-        phase_started,
-    );
-    if formats.keep_wav {
-        let phase_started = Instant::now();
-        std::fs::write(wav_file, &wav_bytes).map_err(|source| ExportPipelineError::Io {
-            path: wav_file.to_path_buf(),
-            source,
-        })?;
-        add_elapsed_phase_ms(
-            &mut output.phase_ms,
-            "post_process.hca.write_wav",
-            phase_started,
-        );
-        output.generated_files.push(wav_file.to_path_buf());
-    }
-    Ok(Some(wav_bytes))
-}
-
-fn encode_hca_mp3(
-    track: &SharedAcbTrack,
-    wav_bytes: Option<&[u8]>,
-    options: &HcaTrackProcessOptions<'_>,
-    output: &mut HcaTrackProcessOutput,
-) -> Result<(), ExportPipelineError> {
-    let mp3 = options.output_dir.join(format!("{}.mp3", track.name));
-    let encode_slot = acquire_media_encode_permit(
-        MediaEncodeKind::Audio,
-        options.audio_encode_concurrency,
-        options.cpu_budget,
-    )?;
-    record_hca_media_encode_acquire(&mut output.phase_ms, &encode_slot);
-    let phase_started = Instant::now();
-    if let Some(wav_bytes) = wav_bytes {
-        convert_wav_bytes_to_mp3_with_backend(
-            wav_bytes,
-            &mp3,
-            options.ffmpeg_path,
-            options.media_backend,
-            options.retry,
-        )?;
-    } else {
-        convert_hca_bytes_to_mp3_with_backend(
-            &track.data,
-            &mp3,
-            options.ffmpeg_path,
-            options.media_backend,
-            options.retry,
-        )?;
-    }
-    drop(encode_slot.cpu_permit);
-    drop(encode_slot.permit);
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.hca.convert_mp3",
-        phase_started,
-    );
-    output.generated_files.push(mp3);
-    Ok(())
-}
-
-fn encode_hca_flac(
-    track: &SharedAcbTrack,
-    wav_bytes: Option<&[u8]>,
-    options: &HcaTrackProcessOptions<'_>,
-    output: &mut HcaTrackProcessOutput,
-) -> Result<(), ExportPipelineError> {
-    let flac = options.output_dir.join(format!("{}.flac", track.name));
-    let encode_slot = acquire_media_encode_permit(
-        MediaEncodeKind::Audio,
-        options.audio_encode_concurrency,
-        options.cpu_budget,
-    )?;
-    record_hca_media_encode_acquire(&mut output.phase_ms, &encode_slot);
-    let phase_started = Instant::now();
-    if let Some(wav_bytes) = wav_bytes {
-        convert_wav_bytes_to_flac_with_backend(
-            wav_bytes,
-            &flac,
-            options.ffmpeg_path,
-            options.media_backend,
-            options.retry,
-        )?;
-    } else {
-        convert_hca_bytes_to_flac_with_backend(
-            &track.data,
-            &flac,
-            options.ffmpeg_path,
-            options.media_backend,
-            options.retry,
-        )?;
-    }
-    drop(encode_slot.cpu_permit);
-    drop(encode_slot.permit);
-    add_elapsed_phase_ms(
-        &mut output.phase_ms,
-        "post_process.hca.convert_flac",
-        phase_started,
-    );
-    output.generated_files.push(flac);
-    Ok(())
 }
